Create sensor _processing .py
The function adds all the numbers in the list using sum (values) and then divides by the number of elements using len (values) to calculate the average
The code uses a for loop to iterate through each sub-list in stations. For each sub-list, it accesses the first element (station[0]) as the station code and the second (station[1]) as the value, printing them in the desired format.
The function loops through each sub-list in stations, unpacks id and pm25, and compares pm25 to the threshold. If it's less, the station is labeled “(safe)”; otherwise, it’s labeled “(danger!)”. The output is formatted using f-strings.